account switching implemented (mostly by moving account related methods from FileActivity.javato DrawerActivity.java)
